+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchListener.java
@@ -1,17 +1,10 @@
 package com.braintreepayments.api;
 
-import android.net.Uri;
-
-import androidx.annotation.Nullable;
-
 public interface BrowserSwitchListener {
     /**
      * The result of a browser switch will be returned in this method.
-     *
-     * @param requestCode the request code used to start this completed request.
-     * @param result The state of the result, one of {@link BrowserSwitchResult#STATUS_SUCCESS} or
-     *     {@link BrowserSwitchResult#STATUS_CANCELED}.
-     * @param returnUri The return uri. {@code null} unless the result is {@link BrowserSwitchResult#STATUS_SUCCESS}.
+     *  @param result The {@link BrowserSwitchResult}. It will have a status of either
+     * {@link BrowserSwitchResult#STATUS_SUCCESS} or {@link BrowserSwitchResult#STATUS_CANCELED}.
      */
-    void onBrowserSwitchResult(int requestCode, BrowserSwitchResult result, @Nullable Uri returnUri);
+    void onBrowserSwitchResult(BrowserSwitchResult result);
 }