+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchResult.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import android.net.Uri;
+
 import androidx.annotation.IntDef;
 
 import org.json.JSONObject;
@@ -9,8 +11,9 @@ import java.lang.annotation.RetentionPolicy;
 
 public class BrowserSwitchResult {
 
-    final private int status;
-    final JSONObject requestMetadata;
+    private final int status;
+    private final Uri deepLinkUri;
+    private final BrowserSwitchRequest request;
 
     @Retention(RetentionPolicy.SOURCE)
     @IntDef({STATUS_SUCCESS, STATUS_CANCELED})
@@ -20,9 +23,15 @@ public class BrowserSwitchResult {
     public static final int STATUS_SUCCESS = 1;
     public static final int STATUS_CANCELED = 2;
 
-    BrowserSwitchResult(@BrowserSwitchStatus int status, JSONObject requestMetadata) {
+    BrowserSwitchResult(@BrowserSwitchStatus int status, BrowserSwitchRequest request) {
+        this(status, request, null);
+    }
+
+    BrowserSwitchResult(@BrowserSwitchStatus int status, BrowserSwitchRequest request, Uri deepLinkUri) {
+        // TODO: unit test
         this.status = status;
-        this.requestMetadata = requestMetadata;
+        this.request = request;
+        this.deepLinkUri = deepLinkUri;
     }
 
     @BrowserSwitchStatus
@@ -31,6 +40,19 @@ public class BrowserSwitchResult {
     }
 
     public JSONObject getRequestMetadata() {
-        return requestMetadata;
+        return request.getMetadata();
+    }
+
+    public int getRequestCode() {
+        return request.getRequestCode();
+    }
+
+    // TODO: determine if this has value for browser switch users
+    public Uri getRequestUri() {
+        return request.getUri();
+    }
+
+    public Uri getDeepLinkUri() {
+        return deepLinkUri;
     }
 }